Use google/jsonapi for easier unmarshal to structs
Using a modified version of github.com/google/jsonapi with one extra
function UnmarshalManyPayloadWithLinks (which is a copy of
UnmarshalManyPayload). The modified function is the same except that it
also returns a map of the pagination links so that the offset can be
parsed and extracted in order to provide easier pagination access to the
user of this package.