Add modified copy of jsonapi.UnmarshalManyPayload
This copy called UnmarshalManyPayloadWithLinks should run the same code
as UnmarshalManyPayload but should also return a map of the pagination
links of the JSON API document.

This commit adds a new file in the google/jsonapi vendored directory
which is not very elegant but it should get the job done till a better
solution is found.

This file needs extra care as the vendor folder might be
destroyed/overwritten by the dep tool (https://github.com/golang/dep).