Use fork instead of google/jsonapi
This commit replaces the usage of google/jsonapi with a fork. The fork
adds an extra function (a copy of UnmarshalManyPayload) which returns
the pagination links of JSON API documents as this functionality is not
provided by google/jsonapi. The need for the links is described in:
https://github.com/nstratos/go-kitsu/issues/2

As a side effect of using the fork, it now seems that there is no longer
a need for vendoring or using a dependency management tool so the vendor
folder, the lock.json and manifest.json files (created by the dep tool)
are deleted.