Add Anime.ShowType & remove enum types
This commit adds a new field ShowType to Anime as well as usage of it in
a test.

It also removes the types AnimeType, MangaType and LibraryEntryStatus.
Since we cannot unmarshal to those types (and we are forced to use
string instead), it seems better to keep them as untyped string
constants to avoid unnecessary conversions.