Add package kitsu/internal/jsonapi
This commit adds a new package jsonapi under kitsu/internal. This new
jsonapi package basically uses the fork of google/jsonapi and isolates
the dependency.

The new jsonapi package is not meant to be exposed to the user of the
kitsu package and since it might change in the future, keeping it under
internal makes sense right now.

Also the code that handles offset is moved in this new package.